Revert "Refactoring: Moved all Views and Models to separate catalogs and separated classes to files with updated propertychanged syntax."
This reverts commit 6f0f9061cdbf4ae099ee3ad146d0be0673f57539.
