+++ BeatSage-Downloader-WPF/AddDownloadWindow.xaml
