+++ BeatSage-Downloader-WPF/AddDownloadWindow.xaml.cs
