+++ BeatSage-Downloader-WPF/App.xaml
@@ -3,7 +3,7 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:BeatSage_Downloader"
              xmlns:properties="clr-namespace:BeatSage_Downloader.Properties"
-             StartupUri="Views/MainWindow.xaml">
+             StartupUri="MainWindow.xaml">
 
     <Application.Resources>
         <ResourceDictionary>
